requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pydantic
pydantic-settings
google-genai
langchain-google-genai
qdrant-client
python-multipart
aiofiles
Pillow
langsmith
boto3
langchain